requirements.txt
pytest
pytest-bdd
pipenv
requests
logger
pytest-html
jsonpath
assertpy
mock
urljoin
jsonschema
testrail
python-dotenv
DateTime
Cerberus
black
autohooks
autohooks-plugin-black
exrex
gcloud
google-cloud-storage
oauth2client
lxml
deepdiff
psycopg2-binary
sshtunnel
boto3
selenium
webdriver_manager
pandas
openpyxl
pytest-reportportal